requirements.txt
streamlit
streamlit_authenticator
PyYAML
bcrypt
pandas
matplotlib
plotly
reportlab
openpyxl
APScheduler
python-dotenv